# server rest/app/src/main/java/org/openapitools/api/InventarioApi.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.22.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package org.openapitools.api;

import org.openapitools.model.InventarioUpdateRequest;
import org.openapitools.model.RespuestaBase;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.ExampleObject;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-05-07T14:57:49.443606900+02:00[Europe/Madrid]", comments = "Generator version: 7.22.0")
@Validated
@Tag(name = "Inventario", description = "Gestión del inventario de vehículos")
public interface InventarioApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    String PATH_INVENTARIO_VEHICULO_ID_GET = "/inventario/{vehiculoId}";
    /**
     * GET /inventario/{vehiculoId} : Consultar inventario del vehículo
     *
     * @param vehiculoId  (required)
     * @param wsKey Clave de acceso requerida para realizar la operación. (required)
     * @return Estado del vehículo obtenido. (status code 200)
     *         or Error en la petición (ej. clave WSKey inválida, recurso no encontrado) (status code 404)
     */
    @Operation(
        operationId = "inventarioVehiculoIdGet",
        summary = "Consultar inventario del vehículo",
        tags = { "Inventario" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Estado del vehículo obtenido.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = RespuestaBase.class))
            }),
            @ApiResponse(responseCode = "404", description = "Error en la petición (ej. clave WSKey inválida, recurso no encontrado)", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = RespuestaBase.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = InventarioApi.PATH_INVENTARIO_VEHICULO_ID_GET,
        produces = { "application/json" }
    )
    default ResponseEntity<RespuestaBase> inventarioVehiculoIdGet(
        @Parameter(name = "vehiculoId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("vehiculoId") String vehiculoId,
        @NotNull @Parameter(name = "WSKey", description = "Clave de acceso requerida para realizar la operación.", required = true, in = ParameterIn.QUERY) @Valid @RequestParam(value = "WSKey", required = true) String wsKey
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"mensaje\" : \"mensaje\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"mensaje\" : \"mensaje\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_INVENTARIO_VEHICULO_ID_PUT = "/inventario/{vehiculoId}";
    /**
     * PUT /inventario/{vehiculoId} : Actualizar estado del vehículo
     *
     * @param vehiculoId  (required)
     * @param wsKey Clave de acceso requerida para realizar la operación. (required)
     * @param inventarioUpdateRequest  (required)
     * @return Inventario actualizado. (status code 200)
     *         or Error en la petición (ej. clave WSKey inválida, recurso no encontrado) (status code 400)
     */
    @Operation(
        operationId = "inventarioVehiculoIdPut",
        summary = "Actualizar estado del vehículo",
        tags = { "Inventario" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Inventario actualizado.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = RespuestaBase.class))
            }),
            @ApiResponse(responseCode = "400", description = "Error en la petición (ej. clave WSKey inválida, recurso no encontrado)", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = RespuestaBase.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = InventarioApi.PATH_INVENTARIO_VEHICULO_ID_PUT,
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    default ResponseEntity<RespuestaBase> inventarioVehiculoIdPut(
        @Parameter(name = "vehiculoId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("vehiculoId") String vehiculoId,
        @NotNull @Parameter(name = "WSKey", description = "Clave de acceso requerida para realizar la operación.", required = true, in = ParameterIn.QUERY) @Valid @RequestParam(value = "WSKey", required = true) String wsKey,
        @Parameter(name = "InventarioUpdateRequest", description = "", required = true) @Valid @RequestBody InventarioUpdateRequest inventarioUpdateRequest
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"mensaje\" : \"mensaje\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"mensaje\" : \"mensaje\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}